Replace AdminResolverInterface with internal methode.
